+++ src/main/kotlin/com/projectcitybuild/core/database/DataSource.kt
@@ -3,12 +3,12 @@ package com.projectcitybuild.core.database
 import co.aikar.idb.DatabaseOptions
 import co.aikar.idb.HikariPooledDatabase
 import co.aikar.idb.PooledDatabaseOptions
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import javax.inject.Singleton
 
 @Singleton
 class DataSource(
-    private val logger: PlatformLogger,
+    private val logger: Logger,
     private val hostName: String,
     private val port: Int = 3306,
     private val databaseName: String,
