+++ src/main/kotlin/com/projectcitybuild/core/database/Migration.kt
@@ -2,7 +2,7 @@ package com.projectcitybuild.core.database
 
 import co.aikar.idb.HikariPooledDatabase
 import com.projectcitybuild.entities.migrations.* // ktlint-disable no-wildcard-imports
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 
 object Migration {
 
@@ -20,7 +20,7 @@ object Migration {
 
     fun executeIfNecessary(
         database: HikariPooledDatabase,
-        logger: PlatformLogger,
+        logger: Logger,
         currentVersion: Int
     ) {
         var version = currentVersion
