+++ src/main/kotlin/com/projectcitybuild/features/utilities/usecases/DataImportUseCase.kt
@@ -5,7 +5,7 @@ import com.projectcitybuild.core.database.DataSource
 import com.projectcitybuild.entities.IPBan
 import com.projectcitybuild.repositories.IPBanRepository
 import com.projectcitybuild.support.spigot.commands.InvalidCommandArgumentsException
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import com.projectcitybuild.support.textcomponent.send
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
@@ -20,7 +20,7 @@ import javax.inject.Inject
 class DataImportUseCase @Inject constructor(
     private val plugin: Plugin,
     private val dataSource: DataSource,
-    private val logger: PlatformLogger,
+    private val logger: Logger,
 ) {
     @Serializable
     data class SpigotBannedIP(
