+++ src/main/kotlin/com/projectcitybuild/features/warps/usecases/TeleportToWarpUseCase.kt
@@ -7,7 +7,7 @@ import com.projectcitybuild.modules.nameguesser.NameGuesser
 import com.projectcitybuild.plugin.events.PlayerPreWarpEvent
 import com.projectcitybuild.repositories.WarpRepository
 import com.projectcitybuild.support.spigot.eventbroadcast.LocalEventBroadcaster
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import org.bukkit.Location
 import org.bukkit.Server
 import org.bukkit.entity.Player
@@ -16,7 +16,7 @@ import javax.inject.Inject
 class TeleportToWarpUseCase @Inject constructor(
     private val warpRepository: WarpRepository,
     private val nameGuesser: NameGuesser,
-    private val logger: PlatformLogger,
+    private val logger: Logger,
     private val localEventBroadcaster: LocalEventBroadcaster,
     private val server: Server,
 ) {
