+++ src/main/kotlin/com/projectcitybuild/modules/errorreporting/adapters/SentryErrorReporter.kt
@@ -3,14 +3,14 @@ package com.projectcitybuild.modules.errorreporting.adapters
 import com.projectcitybuild.modules.config.Config
 import com.projectcitybuild.modules.config.ConfigKeys
 import com.projectcitybuild.modules.errorreporting.ErrorReporter
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import dagger.Reusable
 import io.sentry.Sentry
 
 @Reusable
 class SentryErrorReporter(
     private val config: Config,
-    private val logger: PlatformLogger,
+    private val logger: Logger,
 ) : ErrorReporter {
 
     override fun bootstrap() {
