+++ src/main/kotlin/com/projectcitybuild/plugin/assembly/providers/ErrorReporterProvider.kt
@@ -3,7 +3,7 @@ package com.projectcitybuild.plugin.assembly.providers
 import com.projectcitybuild.modules.config.Config
 import com.projectcitybuild.modules.errorreporting.ErrorReporter
 import com.projectcitybuild.modules.errorreporting.adapters.SentryErrorReporter
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import dagger.Module
 import dagger.Provides
 
@@ -13,7 +13,7 @@ class ErrorReporterProvider {
     @Provides
     fun provideErrorReporter(
         config: Config,
-        logger: PlatformLogger,
+        logger: Logger,
     ): ErrorReporter {
         return SentryErrorReporter(config, logger)
     }
