+++ src/main/kotlin/com/projectcitybuild/plugin/assembly/providers/PermissionsProvider.kt
@@ -2,7 +2,7 @@ package com.projectcitybuild.plugin.assembly.providers
 
 import com.projectcitybuild.modules.permissions.Permissions
 import com.projectcitybuild.modules.permissions.adapters.LuckPermsPermissions
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import dagger.Module
 import dagger.Provides
 import javax.inject.Singleton
@@ -12,7 +12,7 @@ class PermissionsProvider {
 
     @Singleton
     @Provides
-    fun providePermissions(logger: PlatformLogger): Permissions {
+    fun providePermissions(logger: Logger): Permissions {
         return LuckPermsPermissions(logger)
     }
 }
