+++ src/main/kotlin/com/projectcitybuild/plugin/integrations/essentials/EssentialsIntegration.kt
@@ -4,7 +4,7 @@ import com.earth2me.essentials.Essentials
 import com.projectcitybuild.core.SpigotListener
 import com.projectcitybuild.plugin.events.PlayerPreWarpEvent
 import com.projectcitybuild.plugin.integrations.SpigotIntegration
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import dagger.Reusable
 import org.bukkit.event.EventHandler
 import org.bukkit.plugin.Plugin
@@ -13,7 +13,7 @@ import javax.inject.Inject
 @Reusable
 class EssentialsIntegration @Inject constructor(
     private val plugin: Plugin,
-    private val logger: PlatformLogger,
+    private val logger: Logger,
 ) : SpigotListener, SpigotIntegration {
 
     class EssentialsAPINotFoundException : Exception("Essentials plugin not found")
