+++ src/main/kotlin/com/projectcitybuild/plugin/listeners/AsyncPreLoginListener.kt
@@ -4,7 +4,7 @@ import com.projectcitybuild.core.SpigotListener
 import com.projectcitybuild.features.aggregate.ConnectPlayerUseCase
 import com.projectcitybuild.modules.datetime.formatter.DateTimeFormatter
 import com.projectcitybuild.modules.errorreporting.ErrorReporter
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import com.projectcitybuild.support.textcomponent.add
 import kotlinx.coroutines.runBlocking
 import net.md_5.bungee.api.ChatColor
@@ -17,7 +17,7 @@ import javax.inject.Inject
 
 class AsyncPreLoginListener @Inject constructor(
     private val connectPlayerUseCase: ConnectPlayerUseCase,
-    private val logger: PlatformLogger,
+    private val logger: Logger,
     private val dateTimeFormatter: DateTimeFormatter,
     private val errorReporter: ErrorReporter,
 ) : SpigotListener {
