+++ src/main/kotlin/com/projectcitybuild/plugin/listeners/PlayerJoinListener.kt
@@ -4,7 +4,7 @@ import com.projectcitybuild.core.SpigotListener
 import com.projectcitybuild.plugin.events.FirstTimeJoinEvent
 import com.projectcitybuild.repositories.PlayerConfigRepository
 import com.projectcitybuild.support.spigot.eventbroadcast.LocalEventBroadcaster
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import com.projectcitybuild.support.textcomponent.add
 import net.md_5.bungee.api.ChatColor
 import net.md_5.bungee.api.chat.TextComponent
@@ -19,7 +19,7 @@ class PlayerJoinListener @Inject constructor(
     private val server: Server,
     private val localEventBroadcaster: LocalEventBroadcaster,
     private val playerConfigRepository: PlayerConfigRepository,
-    private val logger: PlatformLogger,
+    private val logger: Logger,
 ) : SpigotListener {
 
     @EventHandler
