+++ src/main/kotlin/com/projectcitybuild/repositories/PlayerGroupRepository.kt
@@ -4,7 +4,7 @@ import com.projectcitybuild.core.http.APIRequestFactory
 import com.projectcitybuild.core.http.core.APIClient
 import com.projectcitybuild.modules.config.Config
 import com.projectcitybuild.modules.config.ConfigStorageKey
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import java.util.UUID
 import javax.inject.Inject
 
@@ -12,7 +12,7 @@ class PlayerGroupRepository @Inject constructor(
     private val apiRequestFactory: APIRequestFactory,
     private val apiClient: APIClient,
     private val config: Config,
-    private val logger: PlatformLogger,
+    private val logger: Logger,
 ) {
     class AccountNotLinkedException : Exception()
