+++ src/main/kotlin/com/projectcitybuild/support/spigot/listeners/SpigotListenerRegistry.kt
@@ -1,7 +1,7 @@
 package com.projectcitybuild.support.spigot.listeners
 
 import com.github.shynixn.mccoroutine.bukkit.registerSuspendingEvents
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import dagger.Reusable
 import org.bukkit.event.HandlerList
 import org.bukkit.event.Listener
@@ -11,7 +11,7 @@ import javax.inject.Inject
 @Reusable
 class SpigotListenerRegistry @Inject constructor(
     private val plugin: JavaPlugin,
-    private val logger: PlatformLogger
+    private val logger: Logger
 ) {
     fun register(listener: Listener) {
         plugin.server.pluginManager.registerSuspendingEvents(listener, plugin)
