+++ src/main/kotlin/com/projectcitybuild/support/spigot/scheduler/Scheduler.kt
@@ -2,7 +2,7 @@ package com.projectcitybuild.support.spigot.scheduler
 
 import com.projectcitybuild.core.utilities.AsyncTask
 
-interface PlatformScheduler {
+interface Scheduler {
 
     fun <T> async(task: ((T) -> Unit) -> Unit): AsyncTask<T>
     fun sync(task: () -> Unit)
