+++ src/test/com/projectcitybuild/plugin/listeners/AsyncPreLoginListenerTest.kt
@@ -4,7 +4,7 @@ import com.projectcitybuild.DateTimeFormatterMock
 import com.projectcitybuild.GameBanMock
 import com.projectcitybuild.features.aggregate.ConnectPlayerUseCase
 import com.projectcitybuild.modules.errorreporting.ErrorReporter
-import com.projectcitybuild.support.spigot.logger.PlatformLogger
+import com.projectcitybuild.support.spigot.logger.Logger
 import com.projectcitybuild.stubs.IPBanMock
 import kotlinx.coroutines.test.runTest
 import org.bukkit.event.player.AsyncPlayerPreLoginEvent
@@ -34,7 +34,7 @@ class AsyncPreLoginListenerTest {
 
         listener = AsyncPreLoginListener(
             connectPlayerUseCase = connectPlayerUseCase,
-            mock(PlatformLogger::class.java),
+            mock(Logger::class.java),
             DateTimeFormatterMock(),
             errorReporter,
         )
